Fix point transformation for higher layers
When clicking on a higher layer of an element that's inside a scrollable, the scrollable invalidated clicks outside its boundaries.

This commit adds an optional parameter to convertFor that specifies the layer of the point relative to the layer of the container.
The scrollable now only invalidates clicks outside itself for elements of same or lower layer.
Additionally the order of point transformation in AbstractGuiDropdownMenu has been changed to fix some edge cases.